Rework code to use JSON
Rework the code base to use JSON-formatted text files (with optional
comments, as per Python's commentjson library, available at
https://pypi.python.org/pypi/jsoncomment).

This completely replaces the old ad-hoc text format.

The main driver program 'buv' has been cleaned up and now calls out the
following essential functionality:

  fill-template: Used to fill templates with hashes of files
  automatically. Very useful for testing, but might also have a use in
  production.

  validate: Given a bunch of input JSON files, load all of them and
  figure out the proposals that have been voted on.

The validation checks are still not complete. The format of the  hopefully
shows the way forward on implementing the rest of the validation checks,
though.

The test example, formerly in the testexample1/ directory, has been
reworked so that it is automatically produced from the templates in the
testexample1_template/ directory. A short ad-hoc script named
"make_testexample1.sh" does so.